+++ randovania/games/samus_returns/exporter/patch_data_factory.py
@@ -615,7 +615,7 @@ class MSRPatchDataFactory(PatchDataFactory):
 
     def _objective(self, config: MSRConfiguration) -> dict[str[int], str[str]]:
         return {
-            "required_dna": 39 - (config.artifacts.placed_artifacts - config.artifacts.required_artifacts),
+            "required_dna": config.artifacts.required_artifacts,
             "final_boss": "Ridley",
         }
 
@@ -683,7 +683,7 @@ class MSRPatchDataFactory(PatchDataFactory):
                 "heat": self.configuration.constant_heat_damage,
                 "lava": self.configuration.constant_lava_damage,
             },
-            "objective": self._objective(),
+            "objective": self._objective(self.configuration),
             "layout_uuid": str(self.players_config.get_own_uuid()),
             "enable_remote_lua": self.cosmetic_patches.enable_remote_lua or self.players_config.is_multiworld,
         }