+++ test/games/test_log_file_export.py
@@ -119,6 +119,7 @@ def pytest_generate_tests(metafunc: _pytest.python.Metafunc) -> None:
         "samus_returns/start_inventory.rdvgame",  # test for starting inventory and export ids
         "samus_returns/door_lock.rdvgame",  # starter preset + door lock
         "samus_returns/door_lock_access_open.rdvgame",  # door lock + access open doors
+        "samus_returns/custom_required_dna.rdvgame",  # custom required dna 20/30
         # Super Metroid
         "sdm_test_game.rdvgame",
     ]